+++ dflib/src/main/java/com/nhl/dflib/series/ColumnMappedSeries.java
@@ -62,12 +62,12 @@ public class ColumnMappedSeries<S, T> implements Series<T> {
     }
 
     @Override
-    public Series<T> backFillNulls() {
-        return materialize().backFillNulls();
+    public Series<T> fillNullsBackwards() {
+        return materialize().fillNullsBackwards();
     }
 
     @Override
-    public Series<T> forwardFillNulls() {
-        return materialize().forwardFillNulls();
+    public Series<T> fillNullsForward() {
+        return materialize().fillNullsForward();
     }
 }